refactor: ~Body 클래스들 패키지 변경
