+++ src/test/java/gift/client/KakaoApiClientTest.java
@@ -2,7 +2,7 @@ package gift.client;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gift.dto.body.KakaoTokenRequestBody;
+import gift.client.requestBody.KakaoTokenRequestBody;
 import gift.dto.response.KakaoTokenResponse;
 import gift.exception.KakaoApiHasProblemException;
 import org.junit.jupiter.api.DisplayName;
